tile problem
Generator looks better, but there is a problem that a few different tiles can be on one spot. Furthermore, we can't know exaclty % of the map for each biom.

It can be solved by adding an array of bools with 'trues', in filled positions.
This can even be used to make sure, that every biom is same % of the map.

But ofc it is another pile of math, and we have to remember it must be quick.